singularity: Split into more files.
